[script] Module serialization

+++ src/xci/data/BinaryWriter.cpp
@@ -35,7 +35,7 @@ void BinaryWriter::write_content()
     uint8_t* iter = header + 4;
     assert(is_root_group());
     size_t content_size = group_buffer().size() + (m_crc32 ? 6 : 0);
-    assert(content_size < 0x400'0000'0000LLU);  // up to 4TB
+    assert(uint64_t(content_size) < 0x400'0000'0000LLU);  // up to 4TB
     leb128_encode(iter, content_size);
 
     const size_t header_size = iter - header;
